Modifica modelo basico de vistas. Agrega .env
