--- Models/.gitignore
@@ -1,1 +1,0 @@
-conexion.php
