+++ composer.json
@@ -1,0 +1,13 @@
+{
+    "name": "usuario/compu-visual-p",
+    "authors": [
+        {
+            "name": "ArielParedesLozada",
+            "email": "152462807+ArielParedesLozada@users.noreply.github.com"
+        }
+    ],
+    "require": {
+        "vlucas/phpdotenv": "^5.6",
+        "fpdf/fpdf": "^1.86"
+    }
+}
